requirements.txt

Pillow
easyocr
gradio
openai